Various windows fixes.

+++ client/plugin/src/integTest/groovy/BaseE2EScript.groovy
@@ -1,5 +1,6 @@
-package com.nxt;
+package com.nxt
 
+import com.nxt.config.Util;
 import org.gradle.testkit.runner.GradleRunner
 import spock.lang.Specification
 import spock.util.concurrent.PollingConditions
@@ -8,6 +9,8 @@ import com.nxt.Trouble
 class BaseE2ESpec extends Specification {
     def cleanupSpec() {
         // Kill all the Unity processes we start.
-        "pkill Unity".execute()
+        if (Util.OnOSX()) {
+            "pkill Unity".execute()
+        }
     }
 }
